Move MakepadInputConnection into its own file
